# src/main/resources/api.yaml
openapi: 3.0.3

info:
  description: "YakShop API"
  version: "1.0.0"
  title: YakShop API
  contact:
    email: iggydv12@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'

paths:
  /yak-shop/load:
    post:
      summary: Load a new herd and reset the database
      operationId: loadHerd
      responses:
        '205':
          description: herd loaded successfully
        '400':
          description: invalid input, object invalid
      requestBody:
        description: Herd in XML format to load
        required: true
        content:
          application/xml:
            schema:
              $ref: '#/components/schemas/herdRq'

  /yak-shop/stock/{T}:
    get:
      summary: retrieve total stock at day T
      operationId: getStock
      parameters:
        - name: T
          in: path
          required: true
          description: day
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100000
            example: 5
      responses:
        '200':
          description: stock retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/stock'
        '400':
          description: invalid input
        '404':
          description: object does not exist

  /yak-shop/herd/{T}:
    get:
      summary: retrieve a view of the herd at day T
      operationId: getHerdStatus
      parameters:
        - name: T
          in: path
          required: true
          description: day
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100000
            example: 5
      responses:
        '200':
          description: stock retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/herdStatus'
        '400':
          description: invalid input
        '404':
          description: object does not exist

  /yak-shop/order/{T}:
    post:
      summary: Load a new herd and reset the database
      operationId: order
      parameters:
        - name: T
          in: path
          required: true
          description: day
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100000
            example: 5
      responses:
        '201':
          description: order placed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/stock'
        '206':
          description: only part of the order was acceptd (Deliverable oreder items will be present in the response body)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/stock'
        '404':
          description: full order not in stock.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/order'
        description: order items

components:
  schemas:
    Sex:
      type: string
      xml:
        name: 'sex'
        attribute: true
      enum: [m, f]

    labYakRq:
      type: object
      xml:
        name: 'labyak'
      required:
        - name
        - age
        - sex
      properties:
        name:
          description: yak name
          type: string
          xml:
            name: 'name'
            attribute: true
          minLength: 1
          example: Lilly
        age:
          description: yak age
          type: number
          xml:
            name: 'age'
            attribute: true
          format: double
          minimum: 0.00
          maximum: 10.00
          example: 5.12
        sex:
          description: yak gender
          $ref: '#/components/schemas/Sex'

    labYakRs:
      type: object
      required:
        - name
        - age
        - age-last-shaved
      properties:
        name:
          type: string
          minLength: 1
          example: Lilly
        age:
          description: yak age
          type: number
          format: double
          minimum: 0.00
          maximum: 10.00
          example: 5.12
        age-last-shaved:
          description: the day on which the yak was last shaved
          type: number
          format: double
          minimum: 0.00
          example: 5.12

    herdRq:
      type: array
      required:
        - herd
      xml:
        wrapped: true
      items:
        $ref: '#/components/schemas/labYakRq'

    stock:
      type: object
      required:
        - milk
        - skins
      properties:
        milk:
          description: Total amount of milk in stock
          type: number
          format: double
          minimum: 0.0
          example: 100.00
        skins:
          description: Total amount of skins in stock
          type: integer
          format: int32
          minimum: 0
          example: 1

    order:
      type: object
      required:
        - customer
        - stock
      properties:
        customer:
          type: string
          minLength: 1
          example: YakMan
        stock:
          description: Total amount of skins/milk the customer wishes to order
          $ref: '#/components/schemas/stock'

    herdStatus:
      type: object
      required:
        - herd
      properties:
        herd:
          type: array
          items:
            $ref: '#/components/schemas/labYakRs'